Crate subspace_networking

Networking functionality of Subspace Network, primarily used for DSN (Distributed Storage Network).

Re-exports

pub use libp2p;

Modules

multimess

Structs

BootstrappedNetworkingParameters

Networking manager implementation with bootstrapped addresses. All other operations muted.

Config

Node configuration.

GenericRequestHandler

NetworkingParametersManager

Handles networking parameters. It manages network parameters set and its persistence.

Node

Implementation of a network node on Subspace Network.

NodeRunner

Runner for the Node.

ObjectMappingsRequest

Object-mapping protocol request.

ObjectMappingsResponse

Object-mapping protocol request.

PeerInfo

Defines peer current state.

PeerInfoRequest

Peer-info protocol request.

PeerInfoResponse

Peer-info protocol response.

PiecesByRangeRequest

Pieces-by-range protocol request. Assumes requests with paging.

PiecesByRangeResponse

Pieces-by-range protocol response. Assumes requests with paging.

PiecesToPlot

Collection of pieces that potentially need to be plotted

TopicSubscription

Topic subscription, will unsubscribe when last instance is dropped for a particular topic.

Enums

CircuitRelayClientError

CreationError

Errors that might happen during network creation.

GetClosestPeersError

PeerSyncStatus

Defines peer synchronization status.

RelayMode

Defines relay configuration for the Node

SendRequestError

SubscribeError

Statics

PUB_SUB_ARCHIVING_TOPIC

Traits

GenericRequest

Generic request with associated response

Functions

create

Create a new network node and node runner instances.

Type Definitions

ObjectMappingsRequestHandler

Create a new object-mappings request handler.

PeerInfoRequestHandler

Create a new peer-info request handler.

PiecesByRangeRequestHandler

Create a new pieces-by-range request handler.