subspace_networking::protocols::request_response::handlers
Module cached_piece_by_index

Helper for incoming cached piece requests.

Request handler can be created with CachedPieceByIndexRequestHandler.

Structs

CachedPieceByIndexRequest

Cached-piece-by-index request.

CachedPieceByIndexResponse

Cached-piece-by-index response, may be cached piece or stored in one of the farms

ClosestPeers

Closest peers

Enums

PieceResult

Piece result contains either piece itself or the closest known peers to the piece index

Type Aliases

CachedPieceByIndexRequestHandler

Cached-piece-by-index request handler