Struct subspace_networking::Node

pub struct Node { /* private fields */ }

Implementation of a network node on Subspace Network.

Implementations

impl Node

pub fn id(&self) -> PeerId

Node’s own local ID.

pub fn online_status_observer(&self) -> &Receiver<bool>

pub async fn get_value( &self, key: Multihash ) -> Result<impl Stream<Item = PeerRecord>, GetValueError>

pub async fn put_value( &self, key: Multihash, value: Vec<u8> ) -> Result<impl Stream<Item = ()>, PutValueError>

pub async fn subscribe( &self, topic: Sha256Topic ) -> Result<TopicSubscription, SubscribeError>

pub async fn publish( &self, topic: Sha256Topic, message: Vec<u8> ) -> Result<(), PublishError>

pub async fn send_generic_request<Request>( &self, peer_id: PeerId, request: Request ) -> Result<Request::Response, SendRequestError>where Request: GenericRequest,

Sends the generic request to the peer and awaits the result.

pub async fn get_closest_peers( &self, key: Multihash ) -> Result<impl Stream<Item = PeerId>, GetClosestPeersError>

Get closest peers by multihash key using Kademlia DHT.

pub async fn wait_for_connected_peers( &self ) -> Result<(), CheckConnectedPeersError>

Waits for peers connection to the swarm and for Kademlia address registration.

pub async fn start_announcing( &self, key: Key ) -> Result<impl Stream<Item = ()>, AnnounceError>

Start announcing item by its key. Initiate ‘start_providing’ Kademlia operation.

pub async fn stop_announcing( &self, key: Multihash ) -> Result<(), StopAnnouncingError>

Stop announcing item by its key. Initiate ‘stop_providing’ Kademlia operation.

pub async fn get_providers( &self, key: Multihash ) -> Result<impl Stream<Item = PeerId>, GetProvidersError>

Get item providers by its key. Initiate ‘providers’ Kademlia operation.

pub async fn ban_peer(&self, peer_id: PeerId) -> Result<(), SendError>

Ban peer with specified peer ID.

pub fn listeners(&self) -> Vec<Multiaddr>

Node’s own addresses where it listens for incoming requests.

pub fn on_new_listener( &self, callback: Arc<dyn Fn(&Multiaddr) + Send + Sync + 'static> ) -> HandlerId

Callback is called when node starts listening on new address.

pub fn on_announcement( &self, callback: Arc<dyn Fn(&ProviderRecord, &Arc<InboundStreamEventGuard>) + Send + Sync + 'static> ) -> HandlerId

Callback is called when node starts listening on new address.