Trait subspace_networking::ProviderStorage

pub trait ProviderStorage {
    type ProvidedIter<'a>: Iterator<Item = Cow<'a, ProviderRecord>>
    where
        Self: 'a;

    fn add_provider(&mut self, record: ProviderRecord) -> Result<()>;
    fn providers(&self, key: &Key) -> Vec<ProviderRecord> ;
    fn provided(&self) -> Self::ProvidedIter<'_>;
    fn remove_provider(&mut self, k: &Key, p: &PeerId);
}

Required Associated Types

type ProvidedIter<'a>: Iterator<Item = Cow<'a, ProviderRecord>>
where
Self: 'a

Required Methods

fn add_provider(&mut self, record: ProviderRecord) -> Result<()>

Adds a provider record to the store.

A record store only needs to store a number of provider records for a key corresponding to the replication factor and should store those records whose providers are closest to the key.

fn providers(&self, key: &Key) -> Vec<ProviderRecord>

Gets a copy of the stored provider records for the given key.

fn provided(&self) -> Self::ProvidedIter<'_>

Gets an iterator over all stored provider records for which the node owning the store is itself the provider.

fn remove_provider(&mut self, k: &Key, p: &PeerId)

Removes a provider record from the store.