Crate subspace_networking

Networking functionality of Subspace Network, primarily used for DSN (Distributed Storage Network).

Re-exports

pub use utils::multihash::Multihash;

pub use utils::PeerAddress;

pub use libp2p;

Modules

utils

Miscellaneous utilities for networking.

Structs

Config

Node configuration.

GenericRequestHandler

Defines generic request-response protocol handler.

KeyWrapper

Wrapper data structure that allows to work with keys as Kademlia keys, while not storing 32 bytes if the key itself is smaller, while potentially trading a bit of runtime performance.

KnownPeersManager

Handles networking parameters. It manages network parameters set and its persistence.

KnownPeersManagerConfig

Configuration for KnownPeersManager.

Node

Implementation of a network node on Subspace Network.

NodeRunner

Runner for the Node.

PeerAddressRemovedEvent

Defines the event triggered when the peer address is removed from the permanent storage.

PieceByIndexRequest

Piece-by-hash protocol request.

PieceByIndexResponse

Piece-by-hash protocol response.

SegmentHeaderResponse

Segment header by segment indexes protocol response.

TopicSubscription

Topic subscription, will unsubscribe when last instance is dropped for a particular topic.

UniqueRecordBinaryHeap

Limited-size max binary heap for Kademlia records’ keys.

Enums

CreationError

Errors that might happen during network creation.

GetClosestPeersError

Defines errors for get-closest-peers operation.

KademliaMode

Defines Kademlia mode

KnownPeersManagerPersistenceError

Networking parameters persistence errors.

PeerDiscovered

Represents Kademlia events (RoutablePeer, PendingRoutablePeer, UnroutablePeer).

SegmentHeaderRequest

Segment header by segment indexes protocol request.

SendRequestError

Defines errors for send-request operation.

SubscribeError

Defines errors for subscribe operation.

Traits

GenericRequest

Generic request with associated response

KnownPeersRegistry

Defines operations with the networking parameters.

LocalRecordProvider

Trait to be implemented on providers of local records

Functions

construct

Create a new network node and node runner instances.

peer_id

Converts public key from keypair to PeerId. It serves as the shared PeerId generating algorithm.

Type Aliases

PieceByIndexRequestHandler

Create a new piece-by-hash request handler.

SegmentHeaderBySegmentIndexesRequestHandler

Create a new segment-header-by-segment-indexes request handler.