Struct subspace_networking::Node

pub struct Node { /* private fields */ }

Implementation of a network node on Subspace Network.

Implementations

impl Node

pub fn id(&self) -> PeerId

Node’s own local ID.

pub async fn get_value( &self, key: Multihash ) -> Result<impl Stream<Item = PeerRecord>, GetValueError>

Return a value from the Kademlia network of the DSN.

pub async fn put_value( &self, key: Multihash, value: Vec<u8> ) -> Result<impl Stream<Item = ()>, PutValueError>

Puts a value into the Kademlia network of the DSN.

pub async fn subscribe( &self, topic: Sha256Topic ) -> Result<TopicSubscription, SubscribeError>

Subcribe to some topic on the DSN.

pub async fn publish( &self, topic: Sha256Topic, message: Vec<u8> ) -> Result<(), PublishError>

Subcribe a messgo to some topic on the DSN.

pub async fn send_generic_request<Request>( &self, peer_id: PeerId, request: Request ) -> Result<Request::Response, SendRequestError>where Request: GenericRequest,

Sends the generic request to the peer and awaits the result.

pub async fn get_closest_peers( &self, key: Multihash ) -> Result<impl Stream<Item = PeerId>, GetClosestPeersError>

Get closest peers by multihash key using Kademlia DHT.

pub async fn wait_for_connected_peers( &self, timeout: Duration ) -> Result<(), CheckConnectedPeersError>

Waits for peers connection to the swarm and for Kademlia address registration.

pub async fn start_local_announcing( &self, key: Key ) -> Result<bool, StartLocalAnnouncingError>

Start local announcing item by its key. Saves key to the local storage. It doesn’t affect Kademlia operations.

pub async fn stop_local_announcing( &self, key: Multihash ) -> Result<(), StopLocalAnnouncingError>

Stop local announcing item by its key. Removes key from the local storage.

pub async fn get_providers( &self, key: Multihash ) -> Result<impl Stream<Item = PeerId>, GetProvidersError>

Get item providers by its key. Initiate ‘providers’ Kademlia operation.

pub async fn ban_peer(&self, peer_id: PeerId) -> Result<(), SendError>

Ban peer with specified peer ID.

pub fn listeners(&self) -> Vec<Multiaddr>

Node’s own addresses where it listens for incoming requests.

pub fn external_addresses(&self) -> Vec<Multiaddr>

Node’s own addresses observed remotely.

pub fn on_new_listener( &self, callback: Arc<dyn Fn(&Multiaddr) + Send + Sync + 'static> ) -> HandlerId

Callback is called when node starts listening on new address.

pub fn on_num_established_peer_connections_change( &self, callback: Arc<dyn Fn(&usize) + Send + Sync + 'static> ) -> HandlerId

Callback is called when number of established peer connections changes.

pub async fn connected_peers(&self) -> Result<Vec<PeerId>, ConnectedPeersError>

Returns a collection of currently connected peers.

pub async fn bootstrap(&self) -> Result<(), BootstrapError>

Bootstraps Kademlia network

pub fn on_peer_info( &self, callback: Arc<dyn Fn(&NewPeerInfo) + Send + Sync + 'static> ) -> HandlerId

Callback is called when we receive new crate::peer_info::PeerInfo

pub fn on_disconnected_peer( &self, callback: Arc<dyn Fn(&PeerId) + Send + Sync + 'static> ) -> HandlerId

Callback is called when a peer is disconnected.

pub fn on_connected_peer( &self, callback: Arc<dyn Fn(&PeerId) + Send + Sync + 'static> ) -> HandlerId

Callback is called when a peer is connected.