Trait subspace_networking::LocalRecordProvider

pub trait LocalRecordProvider {
    // Required method
    fn record(&self, key: &RecordKey) -> Option<ProviderRecord>;
}

Trait to be implemented on providers of local records

Required Methods

fn record(&self, key: &RecordKey) -> Option<ProviderRecord>

Gets a provider record for key that is stored locally