Trait subspace_networking::ProviderStorage

pub trait ProviderStorage {
    type ProvidedIter<'a>: Iterator<Item = Cow<'a, ProviderRecord>>
       where Self: 'a;

    // Required methods
    fn add_provider(&self, record: ProviderRecord) -> Result<()>;
    fn providers(&self, key: &Key) -> Vec<ProviderRecord>;
    fn provided(&self) -> Self::ProvidedIter<'_>;
    fn remove_provider(&self, k: &Key, p: &PeerId);
}

A trait for providers storages - wrapper around provider functions of the libp2p RecordStore.

Required Associated Types

type ProvidedIter<'a>: Iterator<Item = Cow<'a, ProviderRecord>> where Self: 'a

Provider record iterator.

Required Methods

fn add_provider(&self, record: ProviderRecord) -> Result<()>

Adds a provider record to the store.

A record store only needs to store a number of provider records for a key corresponding to the replication factor and should store those records whose providers are closest to the key.

fn providers(&self, key: &Key) -> Vec<ProviderRecord>

Gets a copy of the stored provider records for the given key.

fn provided(&self) -> Self::ProvidedIter<'_>

Gets an iterator over all stored provider records for which the node owning the store is itself the provider.

fn remove_provider(&self, k: &Key, p: &PeerId)

Removes a provider record from the store.