Trait subspace_networking::RecordStorage

pub trait RecordStorage<'a> {
    type RecordsIter: Iterator<Item = Cow<'a, Record>>;

    fn get(&'a self, k: &Key) -> Option<Cow<'_, Record>>;
    fn put(&mut self, r: Record) -> Result<()>;
    fn remove(&mut self, k: &Key);
    fn records(&'a self) -> Self::RecordsIter;
}

Required Associated Types

type RecordsIter: Iterator<Item = Cow<'a, Record>>

Required Methods

fn get(&'a self, k: &Key) -> Option<Cow<'_, Record>>

Gets a record from the store, given its key.

fn put(&mut self, r: Record) -> Result<()>

Puts a record into the store.

fn remove(&mut self, k: &Key)

Removes the record with the given key from the store.

fn records(&'a self) -> Self::RecordsIter

Gets an iterator over all (value-) records currently stored.